+++ libraries/GCS_MAVLink/GCS.h
@@ -15,7 +15,10 @@
 #include <AP_Common/Bitmask.h>
 #include <AP_Devo_Telem/AP_Devo_Telem.h>
 #include <RC_Channel/RC_Channel.h>
-#include <AP_Rally/AP_Rally.h>
+
+#include "MissionItemProtocol_Waypoints.h"
+#include "MissionItemProtocol_Rally.h"
+#include "ap_message.h"
 
 #define GCS_DEBUG_SEND_MESSAGE_TIMINGS 0
 
@@ -25,79 +28,6 @@
 #define CHECK_PAYLOAD_SIZE(id) if (comm_get_txspace(chan) < packet_overhead()+MAVLINK_MSG_ID_ ## id ## _LEN) return false
 #define CHECK_PAYLOAD_SIZE2(id) if (!HAVE_PAYLOAD_SPACE(chan, id)) return false
 
-//  GCS Message ID's
-/// NOTE: to ensure we never block on sending MAVLink messages
-/// please keep each MSG_ to a single MAVLink message. If need be
-/// create new MSG_ IDs for additional messages on the same
-/// stream
-enum ap_message : uint8_t {
-    MSG_HEARTBEAT,
-    MSG_ATTITUDE,
-    MSG_LOCATION,
-    MSG_SYS_STATUS,
-    MSG_POWER_STATUS,
-    MSG_MEMINFO,
-    MSG_NAV_CONTROLLER_OUTPUT,
-    MSG_CURRENT_WAYPOINT,
-    MSG_VFR_HUD,
-    MSG_SERVO_OUTPUT_RAW,
-    MSG_RC_CHANNELS,
-    MSG_RC_CHANNELS_RAW,
-    MSG_RAW_IMU,
-    MSG_SCALED_IMU,
-    MSG_SCALED_IMU2,
-    MSG_SCALED_IMU3,
-    MSG_SCALED_PRESSURE,
-    MSG_SCALED_PRESSURE2,
-    MSG_SCALED_PRESSURE3,
-    MSG_SENSOR_OFFSETS,
-    MSG_GPS_RAW,
-    MSG_GPS_RTK,
-    MSG_GPS2_RAW,
-    MSG_GPS2_RTK,
-    MSG_SYSTEM_TIME,
-    MSG_SERVO_OUT,
-    MSG_NEXT_MISSION_REQUEST_WAYPOINTS,
-    MSG_NEXT_MISSION_REQUEST_RALLY,
-    MSG_NEXT_PARAM,
-    MSG_FENCE_STATUS,
-    MSG_AHRS,
-    MSG_SIMSTATE,
-    MSG_AHRS2,
-    MSG_AHRS3,
-    MSG_HWSTATUS,
-    MSG_WIND,
-    MSG_RANGEFINDER,
-    MSG_DISTANCE_SENSOR,
-    MSG_TERRAIN,
-    MSG_BATTERY2,
-    MSG_CAMERA_FEEDBACK,
-    MSG_MOUNT_STATUS,
-    MSG_OPTICAL_FLOW,
-    MSG_GIMBAL_REPORT,
-    MSG_MAG_CAL_PROGRESS,
-    MSG_MAG_CAL_REPORT,
-    MSG_EKF_STATUS_REPORT,
-    MSG_LOCAL_POSITION,
-    MSG_PID_TUNING,
-    MSG_VIBRATION,
-    MSG_RPM,
-    MSG_WHEEL_DISTANCE,
-    MSG_MISSION_ITEM_REACHED,
-    MSG_POSITION_TARGET_GLOBAL_INT,
-    MSG_POSITION_TARGET_LOCAL_NED,
-    MSG_ADSB_VEHICLE,
-    MSG_BATTERY_STATUS,
-    MSG_AOA_SSA,
-    MSG_LANDING,
-    MSG_ESC_TELEMETRY,
-    MSG_ORIGIN,
-    MSG_HOME,
-    MSG_NAMED_FLOAT,
-    MSG_EXTENDED_SYS_STATE,
-    MSG_LAST // MSG_LAST must be the last entry in this enum
-};
-
 // convenience macros for defining which ap_message ids are in which streams:
 #define MAV_STREAM_ENTRY(stream_name)           \
     {                                           \
@@ -106,183 +36,6 @@ enum ap_message : uint8_t {
         ARRAY_SIZE(stream_name ## _msgs)        \
     }
 #define MAV_STREAM_TERMINATOR { (streams)0, nullptr, 0 }
-
-// MissionItemProtocol objects are used for transfering missions from
-// a GCS to ArduPilot and vice-versa.
-//
-// There exists one MissionItemProtocol instance for each of the types
-// of item that might be transfered - e.g. MissionItemProtocol_Rally
-// for rally point uploads.  These objects are static in GCS_MAVLINK
-// and used by all of the backends.
-//
-// While prompting the GCS for items required to complete the mission,
-// a link is stored to the link the MissionItemProtocol should send
-// requests out on, and the "receiving" boolean is true.  In this
-// state downloading of items (and the item count!) is blocked.
-// Starting of uploads (for the same protocol) is also blocked -
-// essentially the GCS uploading a set of items (e.g. a mission) has a
-// mutex over the mission.
-class MissionItemProtocol
-{
-public:
-
-    // note that all of these methods are named after the packet they
-    // are handling; the "mission" part just comes as part of that.
-    void handle_mission_request_list(const class GCS_MAVLINK &link,
-                                     const mavlink_mission_request_list_t &packet,
-                                     const mavlink_message_t &msg);
-    void handle_mission_request_int(const GCS_MAVLINK &link,
-                                    const mavlink_mission_request_int_t &packet,
-                                    const mavlink_message_t &msg);
-    void handle_mission_request(const GCS_MAVLINK &link,
-                                const mavlink_mission_request_t &packet,
-                                const mavlink_message_t &msg);
-
-    void handle_mission_count(class GCS_MAVLINK &link,
-                              const mavlink_mission_count_t &packet,
-                              const mavlink_message_t &msg);
-    void handle_mission_write_partial_list(GCS_MAVLINK &link,
-                                           const mavlink_message_t &msg,
-                                           const mavlink_mission_write_partial_list_t &packet);
-
-    // called on receipt of a MISSION_ITEM or MISSION_ITEM_INT packet;
-    // the former is converted to the latter.
-    void handle_mission_item(const mavlink_message_t &msg,
-                             const mavlink_mission_item_int_t &cmd);
-
-    void handle_mission_clear_all(const GCS_MAVLINK &link,
-                                  const mavlink_message_t &msg);
-
-    void queued_request_send();
-    void update();
-
-    bool active_link_is(const GCS_MAVLINK *_link) const { return _link == link; };
-
-    virtual MAV_MISSION_TYPE mission_type() const = 0;
-
-    bool receiving; // currently sending requests and expecting items
-
-protected:
-
-    GCS_MAVLINK *link; // link currently receiving waypoints on
-
-    // return the ap_message which can be queued to be sent to send a
-    // item request to the GCS:
-    virtual ap_message next_item_ap_message_id() const = 0;
-
-    virtual bool clear_all_items() = 0;
-
-    uint16_t        request_last; // last request index
-
-private:
-
-    virtual void truncate(const mavlink_mission_count_t &packet) = 0;
-
-    uint16_t        request_i; // request index
-
-    // waypoints
-    uint8_t         dest_sysid;  // where to send requests
-    uint8_t         dest_compid; // "
-    uint32_t        timelast_receive_ms;
-    uint32_t        timelast_request_ms;
-    const uint16_t  upload_timeout_ms = 8000;
-
-    // support for GCS getting waypoints etc from us:
-    virtual MAV_MISSION_RESULT get_item(const GCS_MAVLINK &_link,
-                                        const mavlink_message_t &msg,
-                                        const mavlink_mission_request_int_t &packet,
-                                        mavlink_mission_item_int_t &ret_packet) = 0;
-
-    void init_send_requests(GCS_MAVLINK &_link,
-                            const mavlink_message_t &msg,
-                            const int16_t _request_first,
-                            const int16_t _request_last);
-
-    void send_mission_ack(const mavlink_message_t &msg, MAV_MISSION_RESULT result) const;
-    void send_mission_ack(const GCS_MAVLINK &link, const mavlink_message_t &msg, MAV_MISSION_RESULT result) const;
-
-    virtual uint16_t item_count() const = 0;
-    virtual uint16_t max_items() const = 0;
-
-    virtual MAV_MISSION_RESULT replace_item(const mavlink_mission_item_int_t &mission_item_int) = 0;
-    virtual MAV_MISSION_RESULT append_item(const mavlink_mission_item_int_t &mission_item_int) = 0;
-
-    virtual void complete(const GCS_MAVLINK &_link) {};
-    virtual void timeout() {};
-
-    void convert_MISSION_REQUEST_to_MISSION_REQUEST_INT(const mavlink_mission_request_t &request,
-                                                        mavlink_mission_request_int_t &request_int);
-};
-
-class MissionItemProtocol_Waypoints : public MissionItemProtocol {
-public:
-    MissionItemProtocol_Waypoints(AP_Mission &_mission) :
-        mission(_mission) {}
-    void truncate(const mavlink_mission_count_t &packet) override;
-    MAV_MISSION_TYPE mission_type() const override { return MAV_MISSION_TYPE_MISSION; }
-
-    void complete(const GCS_MAVLINK &_link) override;
-    void timeout() override;
-
-protected:
-
-    bool clear_all_items() override WARN_IF_UNUSED;
-
-    ap_message next_item_ap_message_id() const override {
-        return MSG_NEXT_MISSION_REQUEST_WAYPOINTS;
-    }
-
-private:
-    AP_Mission &mission;
-
-    uint16_t item_count() const override { return mission.num_commands(); }
-    uint16_t max_items() const override { return mission.num_commands_max(); }
-
-    MAV_MISSION_RESULT replace_item(const mavlink_mission_item_int_t &) override WARN_IF_UNUSED;
-    MAV_MISSION_RESULT append_item(const mavlink_mission_item_int_t &) override WARN_IF_UNUSED;
-
-    MAV_MISSION_RESULT get_item(const GCS_MAVLINK &_link,
-                                const mavlink_message_t &msg,
-                                const mavlink_mission_request_int_t &packet,
-                                mavlink_mission_item_int_t &ret_packet) override WARN_IF_UNUSED;
-};
-
-class MissionItemProtocol_Rally : public MissionItemProtocol {
-public:
-    MissionItemProtocol_Rally(AP_Rally &_rally) :
-        rally(_rally) {}
-    void truncate(const mavlink_mission_count_t &packet) override;
-    MAV_MISSION_TYPE mission_type() const override { return MAV_MISSION_TYPE_RALLY; }
-
-    void complete(const GCS_MAVLINK &_link) override;
-    void timeout() override;
-
-protected:
-
-    ap_message next_item_ap_message_id() const override {
-        return MSG_NEXT_MISSION_REQUEST_RALLY;
-    }
-    bool clear_all_items() override WARN_IF_UNUSED;
-
-private:
-    AP_Rally &rally;
-
-    uint16_t item_count() const override {
-        return rally.get_rally_total();
-    }
-    uint16_t max_items() const override { return rally.get_rally_max(); }
-
-    MAV_MISSION_RESULT replace_item(const mavlink_mission_item_int_t&) override WARN_IF_UNUSED;
-    MAV_MISSION_RESULT append_item(const mavlink_mission_item_int_t&) override WARN_IF_UNUSED;
-
-    MAV_MISSION_RESULT get_item(const GCS_MAVLINK &_link,
-                                const mavlink_message_t &msg,
-                                const mavlink_mission_request_int_t &packet,
-                                mavlink_mission_item_int_t &ret_packet) override WARN_IF_UNUSED;
-
-    static MAV_MISSION_RESULT convert_MISSION_ITEM_INT_to_RallyLocation(const mavlink_mission_item_int_t &cmd, RallyLocation &ret) WARN_IF_UNUSED;
-
-};
 
 ///
 /// @class	GCS_MAVLINK
