+++ libraries/GCS_MAVLink/GCS_Common.cpp
@@ -180,78 +180,6 @@ GCS_MAVLINK::setup_uart(uint8_t instance)
     }
 }
 
-
-/**
- * @brief Send the next pending waypoint, called from deferred message
- * handling code
- */
-void MissionItemProtocol::queued_request_send()
-{
-    if (!receiving) {
-        return;
-    }
-    if (request_i > request_last) {
-        return;
-    }
-    if (link == nullptr) {
-        AP::internalerror().error(AP_InternalError::error_t::gcs_bad_missionprotocol_link);
-        return;
-    }
-    mavlink_msg_mission_request_send(
-        link->get_chan(),
-        dest_sysid,
-        dest_compid,
-        request_i,
-        mission_type());
-    timelast_request_ms = AP_HAL::millis();
-}
-
-void MissionItemProtocol::update()
-{
-    if (!receiving) {
-        // we don't need to do anything unless we're sending requests
-        return;
-    }
-    if (link == nullptr) {
-        AP::internalerror().error(AP_InternalError::error_t::gcs_bad_missionprotocol_link);
-        return;
-    }
-    // stop waypoint receiving if timeout
-    const uint32_t tnow = AP_HAL::millis();
-    if (tnow - timelast_receive_ms > upload_timeout_ms) {
-        receiving = false;
-        timeout();
-        link = nullptr;
-        return;
-    }
-    // resend request if we haven't gotten one:
-    const uint32_t wp_recv_timeout_ms = 1000U + (link->get_stream_slowdown_ms()*20);
-    if (tnow - timelast_request_ms > wp_recv_timeout_ms) {
-        timelast_request_ms = tnow;
-        link->send_message(next_item_ap_message_id());
-    }
-}
-
-void MissionItemProtocol::send_mission_ack(const mavlink_message_t &msg,
-                                           MAV_MISSION_RESULT result) const
-{
-    if (link == nullptr) {
-        AP::internalerror().error(AP_InternalError::error_t::gcs_bad_missionprotocol_link);
-        return;
-    }
-    send_mission_ack(*link, msg, result);
-}
-void MissionItemProtocol::send_mission_ack(const GCS_MAVLINK &_link,
-                                           const mavlink_message_t &msg,
-                                           MAV_MISSION_RESULT result) const
-{
-    mavlink_msg_mission_ack_send(_link.get_chan(),
-                                 msg.sysid,
-                                 msg.compid,
-                                 result,
-                                 mission_type());
-}
-
 void GCS_MAVLINK::send_meminfo(void)
 {
     unsigned __brkval = 0;
@@ -543,62 +471,6 @@ void GCS_MAVLINK::handle_mission_request_list(const mavlink_message_t &msg)
     prot->handle_mission_request_list(*this, packet, msg);
 }
 
-void MissionItemProtocol::handle_mission_request_list(
-    const GCS_MAVLINK &_link,
-    const mavlink_mission_request_list_t &packet,
-    const mavlink_message_t &msg)
-{
-    if (receiving) {
-        // someone is uploading a mission; reject fetching of points
-        // until done or timeout
-        send_mission_ack(_link, msg, MAV_MISSION_DENIED);
-        return;
-    }
-
-    // reply with number of commands in the mission.  The GCS will
-    // then request each command separately
-    mavlink_msg_mission_count_send(_link.get_chan(),
-                                   msg.sysid,
-                                   msg.compid,
-                                   item_count(),
-                                   mission_type());
-}
-
-void MissionItemProtocol::handle_mission_request_int(const GCS_MAVLINK &_link,
-                                                     const mavlink_mission_request_int_t &packet,
-                                                     const mavlink_message_t &msg)
-{
-    if (receiving) {
-        // someone is uploading a mission; reject fetching of points
-        // until done or timeout
-        send_mission_ack(_link, msg, MAV_MISSION_DENIED);
-        return;
-    }
-
-    mavlink_mission_item_int_t ret_packet{};
-
-    ret_packet.target_system = msg.sysid;
-    ret_packet.target_component = msg.compid;
-    ret_packet.seq = packet.seq;
-    ret_packet.mission_type = packet.mission_type;
-
-    const MAV_MISSION_RESULT result_code = get_item(_link, msg, packet, ret_packet);
-
-    if (result_code != MAV_MISSION_ACCEPTED) {
-        // send failure message
-        send_mission_ack(_link, msg, result_code);
-        return;
-    }
-
-    // we already have a filled structure, use it in place of _send:
-    _mav_finalize_message_chan_send(_link.get_chan(),
-                                    MAVLINK_MSG_ID_MISSION_ITEM_INT,
-                                    (const char *)&ret_packet,
-                                    MAVLINK_MSG_ID_MISSION_ITEM_MIN_LEN,
-                                    MAVLINK_MSG_ID_MISSION_ITEM_INT_LEN,
-                                    MAVLINK_MSG_ID_MISSION_ITEM_INT_CRC);
-}
-
 /*
   handle a MISSION_REQUEST mavlink packet
  */
@@ -615,48 +487,6 @@ void GCS_MAVLINK::handle_mission_request_int(const mavlink_message_t &msg)
         prot->handle_mission_request_int(*this, packet, msg);
 }
 
-MAV_MISSION_RESULT MissionItemProtocol_Waypoints::get_item(const GCS_MAVLINK &_link,
-                                                           const mavlink_message_t &msg,
-                                                           const mavlink_mission_request_int_t &packet,
-                                                           mavlink_mission_item_int_t &ret_packet)
-{
-    if (packet.seq != 0 && // always allow HOME to be read
-        packet.seq >= mission.num_commands()) {
-        // try to educate the GCS on the actual size of the mission:
-        mavlink_msg_mission_count_send(_link.get_chan(),
-                                       msg.sysid,
-                                       msg.compid,
-                                       mission.num_commands(),
-                                       MAV_MISSION_TYPE_MISSION);
-        return MAV_MISSION_ERROR;
-    }
-
-    AP_Mission::Mission_Command cmd;
-
-        // retrieve mission from eeprom
-        if (!mission.read_cmd_from_storage(packet.seq, cmd)) {
-            return MAV_MISSION_ERROR;
-        }
-
-        if (!AP_Mission::mission_cmd_to_mavlink_int(cmd, ret_packet)) {
-            return MAV_MISSION_ERROR;
-        }
-
-        // set packet's current field to 1 if this is the command being executed
-        if (cmd.id == (uint16_t)mission.get_current_nav_cmd().index) {
-            ret_packet.current = 1;
-        } else {
-            ret_packet.current = 0;
-        }
-
-        // set auto continue to 1
-        ret_packet.autocontinue = 1;     // 1 (true), 0 (false)
-
-        ret_packet.command = cmd.id;
-
-        return MAV_MISSION_ACCEPTED;
-}
-
 void GCS_MAVLINK::handle_mission_request(const mavlink_message_t &msg)
 {
         // decode
@@ -668,46 +498,6 @@ void GCS_MAVLINK::handle_mission_request(const mavlink_message_t &msg)
             return;
         }
         prot->handle_mission_request(*this, packet, msg);
-}
-
-void MissionItemProtocol::convert_MISSION_REQUEST_to_MISSION_REQUEST_INT(const mavlink_mission_request_t &request, mavlink_mission_request_int_t &request_int)
-{
-    request_int.target_system = request.target_system;
-    request_int.target_component = request.target_component;
-    request_int.seq = request.seq;
-    request_int.mission_type = request.mission_type;
-}
-
-void MissionItemProtocol::handle_mission_request(const GCS_MAVLINK &_link,
-                                                 const mavlink_mission_request_t &packet,
-                                                 const mavlink_message_t &msg
-)
-{
-    mavlink_mission_request_int_t request_int;
-    convert_MISSION_REQUEST_to_MISSION_REQUEST_INT(packet, request_int);
-
-    mavlink_mission_item_int_t item_int{};
-    item_int.target_system = msg.sysid;
-    item_int.target_component = msg.compid;
-
-    MAV_MISSION_RESULT ret = get_item(_link, msg, request_int, item_int);
-    if (ret != MAV_MISSION_ACCEPTED) {
-        return;
-    }
-
-    mavlink_mission_item_t ret_packet{};
-    ret = AP_Mission::convert_MISSION_ITEM_INT_to_MISSION_ITEM(item_int, ret_packet);
-    if (ret != MAV_MISSION_ACCEPTED) {
-        return;
-    }
-
-    // we already have a filled structure, use it in place of _send:
-    _mav_finalize_message_chan_send(_link.get_chan(),
-                                    MAVLINK_MSG_ID_MISSION_ITEM,
-                                    (const char *)&ret_packet,
-                                    MAVLINK_MSG_ID_MISSION_ITEM_MIN_LEN,
-                                    MAVLINK_MSG_ID_MISSION_ITEM_LEN,
-                                    MAVLINK_MSG_ID_MISSION_ITEM_CRC);
 }
 
 /*
@@ -747,66 +537,6 @@ void GCS_MAVLINK::handle_mission_count(const mavlink_message_t &msg)
     prot->handle_mission_count(*this, packet, msg);
 }
 
-void MissionItemProtocol::init_send_requests(GCS_MAVLINK &_link,
-                                             const mavlink_message_t &msg,
-                                             const int16_t _request_first,
-                                             const int16_t _request_last)
-{
-    // set variables to help handle the expected receiving of commands from the GCS
-    timelast_receive_ms = AP_HAL::millis();    // set time we last received commands to now
-    receiving = true;              // record that we expect to receive commands
-    request_i = _request_first;                 // reset the next expected command number to zero
-    request_last = _request_last;         // record how many commands we expect to receive
-
-    dest_sysid = msg.sysid;       // record system id of GCS who wants to upload the mission
-    dest_compid = msg.compid;     // record component id of GCS who wants to upload the mission
-
-    link = &_link;
-
-    timelast_request_ms = AP_HAL::millis();
-    link->send_message(next_item_ap_message_id());
-}
-
-void MissionItemProtocol::handle_mission_count(
-    GCS_MAVLINK &_link,
-    const mavlink_mission_count_t &packet,
-    const mavlink_message_t &msg)
-{
-    if (receiving) {
-        // someone is already uploading a mission.  If we are
-        // receiving from someone then we will allow them to restart -
-        // otherwise we deny.
-        if (msg.sysid != dest_sysid || msg.compid != dest_compid) {
-            // reject another upload until
-            send_mission_ack(_link, msg, MAV_MISSION_DENIED);
-            return;
-        }
-    }
-
-    if (packet.count > max_items()) {
-        send_mission_ack(_link, msg, MAV_MISSION_NO_SPACE);
-        return;
-    }
-
-    truncate(packet);
-
-    if (packet.count == 0) {
-        // no requests to send...
-        send_mission_ack(_link, msg, MAV_MISSION_ACCEPTED);
-        complete(_link);
-        return;
-    }
-
-    // start waypoint receiving
-    init_send_requests(_link, msg, 0, packet.count-1);
-}
-
-void MissionItemProtocol_Waypoints::truncate(const mavlink_mission_count_t &packet)
-{
-    // new mission arriving, truncate mission to be the same length
-    mission.truncate(packet.count);
-}
-
 /*
   handle a MISSION_CLEAR_ALL mavlink packet
  */
@@ -824,26 +554,6 @@ void GCS_MAVLINK::handle_mission_clear_all(const mavlink_message_t &msg)
     }
 
     prot->handle_mission_clear_all(*this, msg);
-}
-
-bool MissionItemProtocol_Waypoints::clear_all_items()
-{
-    return mission.clear();
-}
-
-bool MissionItemProtocol_Rally::clear_all_items()
-{
-    rally.truncate(0);
-    return true;
-}
-
-void MissionItemProtocol::handle_mission_clear_all(const GCS_MAVLINK &_link,
-                                                   const mavlink_message_t &msg)
-{
-    bool success = true;
-    success = success && !receiving;
-    success = success && clear_all_items();
-    send_mission_ack(_link, msg, success ? MAV_MISSION_ACCEPTED : MAV_MISSION_ERROR);
 }
 
 bool GCS_MAVLINK::requesting_mission_items() const
@@ -870,24 +580,6 @@ void GCS_MAVLINK::handle_mission_write_partial_list(const mavlink_message_t &msg
     }
     use_prot->handle_mission_write_partial_list(*this, msg, packet);
 }
-
-void MissionItemProtocol::handle_mission_write_partial_list(GCS_MAVLINK &_link,
-                                                            const mavlink_message_t &msg,
-                                                            const mavlink_mission_write_partial_list_t &packet)
-{
-
-    // start waypoint receiving
-    if ((unsigned)packet.start_index > item_count() ||
-        (unsigned)packet.end_index > item_count() ||
-        packet.end_index < packet.start_index) {
-        gcs().send_text(MAV_SEVERITY_WARNING,"Flight plan update rejected"); // FIXME: Remove this anytime after 2020-01-22
-        send_mission_ack(_link, msg, MAV_MISSION_ERROR);
-        return;
-    }
-
-    init_send_requests(_link, msg, packet.start_index, packet.end_index);
-}
-
 
 /*
   pass mavlink messages to the AP_Mount singleton
@@ -1029,118 +721,6 @@ void GCS_MAVLINK::handle_mission_item(const mavlink_message_t &msg)
     }
 
     prot->handle_mission_item(msg, packet);
-}
-
-MAV_MISSION_RESULT MissionItemProtocol_Waypoints::replace_item(const mavlink_mission_item_int_t &mission_item_int)
-{
-    AP_Mission::Mission_Command cmd;
-
-    const MAV_MISSION_RESULT res = AP_Mission::mavlink_int_to_mission_cmd(mission_item_int, cmd);
-    if (res != MAV_MISSION_ACCEPTED) {
-        return res;
-    }
-
-    // sanity check for DO_JUMP command
-    if (cmd.id == MAV_CMD_DO_JUMP) {
-        if ((cmd.content.jump.target >= item_count() && cmd.content.jump.target > request_last) || cmd.content.jump.target == 0) {
-            return MAV_MISSION_ERROR;
-        }
-    }
-    if (!mission.replace_cmd(cmd.index, cmd)) {
-        return MAV_MISSION_ERROR;
-    }
-    return MAV_MISSION_ACCEPTED;
-}
-
-MAV_MISSION_RESULT MissionItemProtocol_Waypoints::append_item(const mavlink_mission_item_int_t &mission_item_int)
-{
-    // sanity check for DO_JUMP command
-    AP_Mission::Mission_Command cmd;
-
-    const MAV_MISSION_RESULT res = AP_Mission::mavlink_int_to_mission_cmd(mission_item_int, cmd);
-    if (res != MAV_MISSION_ACCEPTED) {
-        return res;
-    }
-
-    if (cmd.id == MAV_CMD_DO_JUMP) {
-        if ((cmd.content.jump.target >= item_count() && cmd.content.jump.target > request_last) || cmd.content.jump.target == 0) {
-            return MAV_MISSION_ERROR;
-        }
-    }
-
-    if (!mission.add_cmd(cmd)) {
-        return MAV_MISSION_ERROR;
-    }
-    return MAV_MISSION_ACCEPTED;
-}
-
-void MissionItemProtocol_Waypoints::complete(const GCS_MAVLINK &_link)
-{
-    _link.send_text(MAV_SEVERITY_INFO, "Flight plan received");
-    AP::logger().Write_EntireMission();
-}
-void MissionItemProtocol_Waypoints::timeout()
-{
-    link->send_text(MAV_SEVERITY_WARNING, "Mission upload timeout");
-}
-
-void MissionItemProtocol::handle_mission_item(const mavlink_message_t &msg, const mavlink_mission_item_int_t &cmd)
-{
-    if (link == nullptr) {
-        AP::internalerror().error(AP_InternalError::error_t::gcs_bad_missionprotocol_link);
-        return;
-    }
-
-    // check if this is the requested waypoint
-    if (cmd.seq != request_i) {
-        send_mission_ack(msg, MAV_MISSION_INVALID_SEQUENCE);
-        return;
-    }
-    // make sure the item is coming from the system that initiated the upload
-    if (msg.sysid != dest_sysid) {
-        send_mission_ack(msg, MAV_MISSION_DENIED);
-        return;
-    }
-    if (msg.compid != dest_compid) {
-        send_mission_ack(msg, MAV_MISSION_DENIED);
-        return;
-    }
-
-    const uint16_t _item_count = item_count();
-
-    MAV_MISSION_RESULT result;
-    if (cmd.seq < _item_count) {
-        // command index is within the existing list, replace the command
-        result = replace_item(cmd);
-    } else if (cmd.seq == _item_count) {
-        // command is at the end of command list, add the command
-        result = append_item(cmd);
-    } else {
-        // beyond the end of the command list, return an error
-        result = MAV_MISSION_ERROR;
-    }
-    if (result != MAV_MISSION_ACCEPTED) {
-        send_mission_ack(msg, result);
-        return;
-    }
-
-    // update waypoint receiving state machine
-    timelast_receive_ms = AP_HAL::millis();
-    request_i++;
-
-    if (request_i > request_last) {
-        send_mission_ack(msg, MAV_MISSION_ACCEPTED);
-        complete(*link);
-        receiving = false;
-        link = nullptr;
-        return;
-    }
-    // if we have enough space, then send the next WP request immediately
-    if (HAVE_PAYLOAD_SPACE(link->get_chan(), MISSION_REQUEST)) {
-        queued_request_send();
-    } else {
-        link->send_message(next_item_ap_message_id());
-    }
 }
 
 ap_message GCS_MAVLINK::mavlink_id_to_ap_message_id(const uint32_t mavlink_id) const
