+++ libraries/GCS_MAVLink/GCS_Rally.cpp
@@ -1,5 +1,7 @@
 /*
-   GCS MAVLink functions related to rally points
+   GCS MAVLink functions related to upload and download of rally
+   points with the ArduPilot-specific protocol comprised of
+   MAVLINK_MSG_ID_RALLY_POINT and MAVLINK_MSG_ID_RALLY_FETCH_POINT.
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -100,86 +102,4 @@ void GCS_MAVLINK::handle_common_rally_message(const mavlink_message_t &msg)
 #endif
         break;
     }
-}
-
-MAV_MISSION_RESULT MissionItemProtocol_Rally::get_item(const GCS_MAVLINK &_link,
-                                                       const mavlink_message_t &msg,
-                                                       const mavlink_mission_request_int_t &packet,
-                                                       mavlink_mission_item_int_t &ret_packet)
-{
-    RallyLocation rallypoint;
-
-    if (!rally.get_rally_point_with_index(packet.seq, rallypoint)) {
-        return MAV_MISSION_INVALID_SEQUENCE;
-    }
-
-    ret_packet.frame = MAV_FRAME_GLOBAL_RELATIVE_ALT_INT;
-    ret_packet.command = MAV_CMD_NAV_RALLY_POINT;
-    ret_packet.x = rallypoint.lat;
-    ret_packet.y = rallypoint.lng;
-    ret_packet.z = rallypoint.alt;
-
-    return MAV_MISSION_ACCEPTED;
-}
-
-void MissionItemProtocol_Rally::truncate(const mavlink_mission_count_t &packet)
-{
-    rally.truncate(packet.count);
-}
-void MissionItemProtocol_Rally::timeout()
-{
-    link->send_text(MAV_SEVERITY_WARNING, "Rally upload timeout");
-}
-
-MAV_MISSION_RESULT MissionItemProtocol_Rally::convert_MISSION_ITEM_INT_to_RallyLocation(const mavlink_mission_item_int_t &cmd, RallyLocation &ret)
-{
-    if (cmd.frame != MAV_FRAME_GLOBAL_RELATIVE_ALT_INT &&
-        cmd.frame != MAV_FRAME_GLOBAL_RELATIVE_ALT) {
-        return MAV_MISSION_UNSUPPORTED_FRAME;
-    }
-    if (!check_lat(cmd.x)) {
-        return MAV_MISSION_INVALID_PARAM5_X;
-    }
-    if (!check_lng(cmd.y)) {
-        return MAV_MISSION_INVALID_PARAM6_Y;
-    }
-    if (cmd.z < INT16_MIN || cmd.z > INT16_MAX) {
-        return MAV_MISSION_INVALID_PARAM7;
-    }
-    ret.lat = cmd.x;
-    ret.lng = cmd.y;
-    ret.alt = cmd.z;
-    return MAV_MISSION_ACCEPTED;
-}
-
-MAV_MISSION_RESULT MissionItemProtocol_Rally::replace_item(const mavlink_mission_item_int_t &cmd)
-{
-    RallyLocation rallyloc;
-    const MAV_MISSION_RESULT ret = convert_MISSION_ITEM_INT_to_RallyLocation(cmd, rallyloc);
-    if (ret != MAV_MISSION_ACCEPTED) {
-        return ret;
-    }
-    if (!rally.set_rally_point_with_index(cmd.seq, rallyloc)) {
-        return MAV_MISSION_ERROR;
-    }
-    return MAV_MISSION_ACCEPTED;
-}
-
-MAV_MISSION_RESULT MissionItemProtocol_Rally::append_item(const mavlink_mission_item_int_t &cmd)
-{
-    RallyLocation rallyloc;
-    const MAV_MISSION_RESULT ret = convert_MISSION_ITEM_INT_to_RallyLocation(cmd, rallyloc);
-    if (ret != MAV_MISSION_ACCEPTED) {
-        return ret;
-    }
-    if (!rally.append(rallyloc)) {
-        return MAV_MISSION_ERROR;
-    }
-    return MAV_MISSION_ACCEPTED;
-}
-
-void MissionItemProtocol_Rally::complete(const GCS_MAVLINK &_link)
-{
-    _link.send_text(MAV_SEVERITY_INFO, "Rally points received");
-    AP::logger().Write_Rally();
 }
